Goodbye react-hook-form, hello own implementation
At first, react-hook-form was good. Very easy to use, a little bit
weird here and there, but overall good. Then the quirks started
arriving. Objects that were copied incorrectly, arrays that were forced
to contain objects, default values that needed to be propagated all over
the place even though they were right there in the form data. In its
quest to avoid unnecessary re-renders, react-hook-form introduced so
much quirky behaviour that can't easily be fixed without breaking
changes.

The worst is probably `useFieldArray` and the shambolic way in which it
handles default values. So much mess.

I wanted to avoid writing my own form handling for the longest time,
since it was nice and convenient to outsource it. Unfortunately the many
weirdnesses of react-hook-form make it unsuitable for further
development. As such, I bit the bullet and rolled my own.

It seems to work well!

Since I specialised it for Condict's needs, I was able to get rid of
several quirky and pointless things. Cleaner code ftw!