app: Rich-text adjustments
You can now customise *all* the block element tags in `formatBlocks()`.
The `<RichContent>` component still only allows headings to be
customised. The motivation behind this: inside `<Card>`s and
`<LinkCard>`s, while we do want headings to be rendered, we *don't* want
them to have heading semantics, as they kinda mess up the heading
structure for screen readers.

Additionally, I have (once again) tweaked the heading sizes. A
difference of 2px in font-size between the various levels was basically
imperceptible. In context, it was very hard to distinguish between <h3>
and <h4>, even when placed right next to each other. As a result of this
change, the lowest officially supported heading level is h4. In
practice, h5 and h6 are too low to ever be used, so away they go.