Replace Inter with Noto Sans
Inter is a very nice and friendly font that works beautifully for UI
text. It is highly legible with a good x-height and lot sof nice
OpenType features. Unfortunately, it is also basically limited to Latin,
Greek and Cyrillic. For an application that's all about language, this
is unacceptable. It doesn't even have full IPA support; some glyphs are
forced to use a fallback font, which looks out of place and ugly. Plus,
while development on Inter hasn't been abandoned by any means, it seems
to have slowed, and various outstanding issues (mostly around Cyrillic)
have not yet been fixed.

The Noto family of fonts has support for loads of languages, with a
harmonious style and extremely high legibility. Currently I've only
bundled the basic Noto Sans (Latin, Greek, Cyrillic), but in future I
may add other Noto families for broader language support.

Sorry to let you go, Inter. It was good for a while.