app: Simplify page data management with specialised hook
The `usePageData` hook takes care of fetching page data and reloading,
updating the page title to fit the data, and managing focus in response
to data changes. I noticed these tasks are required by basically all
pages, so turned it into a special-purpose hook that hides much of the
complexity and reduces repetition.