app: Use less brittle focus management in sidebar tab list
I have noticed that keyboard focus in the sidebar tab list doesn't
always follow the current tab. On rare occasion, it stays on whatever
tab it was on when you focused the tab list. I have not found a reliable
means of reproducing this bug, but I *suspect* it has to do with the
order of focus and blur events. The DOM is messy, and sometimes blur
events fire *after* focus events in a confusing fashion, and I believe
it may have resulted in `hasFocus` becoming false when the tab list in
fact contained the focus.

To resolve this issue, we instead take a ref to the tab list and check
whether the focus is inside it. This seems to work perfectly
consistently and reliably. Yay!