server: Hardcode Unicode collation name
Previously, the collation name was passed in from JS. The way I stored
the name was actually incorrect and memory unsafe, as the `const char*`
obtained from the `Utf8Value` is only valid as long as the `Utf8Value`
is in scope. By sheer happenstance, it all worked out, but it was wrong.
Since I only ever register a single collation that is always named
`unicode`, I decided to just hardcode the name instead.