rte: Reduce unnecessary re-renders
The `BaseEditor` component is now deconstructed into `BaseEditable`,
`EditorContainer` and `EditorToolbar`. The previous `BaseEditor`
component did a little bit too much, which effectively meant it had to
re-render a *lot*. The most important effect of this change is that I've
finally fixed a vexing bug: when clicking in a rich-text editor, the
selection would sometimes, unpredictably, fail to end up inside the
editor. It just sort of disappeared. It had *something* to do with
rendering during the focus event, but I could never figure out what the
exact cause was. Simply reducing re-renders solves the bug.

... unfortunately it appears to have introduced a new bug, where
`editor.focusWithSelection` does not always select. Always more to do.