Improve definition term editing behaviour
To avoid breaking links, we now retain lemma IDs whenever possible,
instead of always creating a new lemma ID when the definition term is
edited to a non-existent lemma. The code contains a comment that
describes this behaviour.

In addition, when editing a definition from a lemma page, the app now
shows the updated lemma - either by opening the new lemma in a new tab,
or by replacing the current tab.