Make font size and line height configurable
There is not merely one font size and one line height, but many that
must be coordinated. For that reason, the settings are limited to a
small set of hardcoded options, each of which is carefully selected to
look decent.

Something I did not do in this commit, which I would like to do at some
point, is to adjust certain dimensions in response to font size changes.
At size 18, the sidebar feels rather narrow, as does the main content
area. It would be nice if these could grow, within reasonable bounds,
as the font size increases.