ui: Remove label prop from Button, Checkbox, Radio, Switch
The `label` prop did one of two things depending on the presence of
children:

- if the component had no children, then the label was the *visible*
  content of the component (always text);
- if the component did have children, then the label became the ARIA
  label (invisible except to assistive technologies).

It has long felt weird to me that the *visible* contents of these
component could be specified in two ways. In practice, also, (excluding
the component test pages) I mostly used either children with optional
icon + `<Localized>`, or an icon child + `label`.

It feels better for the label to be `aria-label`, which more clearly
signals its intent, and for the visible content to be always in the
children.

Note: For toolbar items, the `title` prop is automatically used as the
`aria-label` if the latter is absent. Toolbar buttons often contain
icons, and being able to assign a tooltip and label in one go is
convenient.

Note: Menu items are unaffected. For one, menu item children are the
contents of the item's submenu. Secondly, the menu item label must be a
string so that we can find it when the user types in the menu. Press 'S'
and we jump to (possibly activate) any menu item that starts with 'S'.