ui/shortcut: Replace 'Secondary' with 'Ctrl'
The secondary modifier relies on the modifier state 'OS', which is deprecated. Users should not bind application-level shortcuts to that modifier anyway, and it may become impossible to detect in future versions of Chromium. Hence, we stop relying on it. The contrast between primary and secondary still makes sense for macOS, but now we call it Primary (ctrl or cmd) and Ctrl (always ctrl) instead.