Fix Rollup config
The TypeScript plugin has some strange behaviour that puts type
declarations in the wrong place when an output has a `file` instead of
`dir`, but I managed to find a workaround. It's not very nice, but...

I also move the `.tsbuildinfo` files to `.buildcache`. They don't really
belong in `dist`.