Split server into server and http-server
The `@condict/server` package is now completely independent of Apollo,
meaning we don't have to bundle it with the app (which wouldn't use it
anyway), nor force downstream users to use it. All HTTP functionality is
now in its own package, `@condict/http-server`, a standalone HTTP server
for Condict.

For this commit, I did not simply move all the HTTP functionality over.
I took the opportunity to upgrade to the latest Apollo version, and make
a few API changes to streamline and reduce risk of errors. It seems to
work just fine!

The old error types we relied on are now implemented in
`@condict/server`, as they have been removed from Apollo.

Also, config loading is now *absent* in `@condict/server`. It doesn't
care where the config comes from, as long as the object has the right
shape. For convenience, we still expose functions for validating config
values. The HTTP server now contains the config loading logic.