server: Decouple inflection tables from parts of speech
The original idea was as follows: only verbs can be conjugated like
verbs, so it doesn't make sense to show verbal inflection tables for
nouns. Likewise, verbs can't receive nominal declension, so why would
you make it possible to add such tables to a verb? Similar arguments
apply to other parts of speech.

In practice, however, languages are not so straightforward. There may be
languages in which adjectives can behave like verbs or nouns depending
on context, so it makes sense for them to both conjugate and decline.
Now the inability to add verb tables to an adjective potentially forces
you to duplicate tables in your dictionary.

By making inflection tables a top-level resource under languages, we do
end up shifting some of the burden onto the user, who now has to
structure things themself, but we also give them the power to do as they
please. I believe this change to be ultimately beneficial.

---

I also slid in a few other, minor changes to the database schema, mainly
adding `without rowid` to some tables.