Use 'declare const' for "nominal type" symbols
Symbols used only in the declaration of types such as `IdOf`, where the
symbol is not actually present in the value but is only used for
discrimination or to look up type metadata, do not need to have a
declared value. We can use `declare const` with the type `unique symbol`
to *guarantee* that no code is emitted for them.