ui/Select: Remove options as children, make generic over value type
The way `<Select>` actually ended up being used, we generally want to
preserve type information for the values in the dropdown. For instance,
in the app we have a `<Select>` over a list of `PartOfSpeechId`s and a
`null` value. The previous implementation required us to maintain our
own mapping between string values and `PartOfSpeechId | null`, in both
directions. With the new implementation, you provide a `value` as part
of the `SelectOption` and everything else Just Works; the translation
happens automagically. Basically, the code gets nice cleaner.