app: Rework data hooks
Data fetching occurs in many parts of the app. Sometimes we want to
fetch data once (e.g. for an edit panel), sometimes we want it to update
in response to dictionary events, sometimes we even have initial data
and *only* want to update on certain events. That last case had several
existing and subtly different implementations. In the future, I intend
to add support for automatic retries in response to recoverable errors
and other nice features, and having many data fetching implementations
drastically complicates that.

This commit splits `useData` into three separate hooks:

- `useData`: Fetch once or when variables change, no dynamic refetching
  in response to events. Used by edit panels to load initial data. Skips
  all the machinery related to dictionary events and the like.
- `useLiveData`: Fetch once or seed with initial data, and dynamically
  refetch in response to events. Event predicate is not optional.
- `useLiveMutData`: Same as the above, but also returns a callback for
  updating the locally cached result. This particular hook is intended
  for use cases where we want to update a list of things in response to
  a user action, while still allowing the list to be refetched behind
  the scenes as events come in.

Moreover, it is sometimes desirable to manipulate the returned data in
some way, which often requires `useMemo` plus various conditions. To
better support this use case, all three hooks accept a `mapData` that is
called when data is received. If initial data is supplied, it must match
the return type of `mapData`.

And lastly, errors. Previously, the result of a query was always either
'loading' or 'data', where the latter includes any errors that may have
occurred. Now we have a third state: 'error'. If any error occurs at
all, all of the hooks bail immediately. Partial data is not acceptable.
This requires different handling in a few places.