app: Rework translations slightly
Previously, the renderer only held the current bundle and the default
bundle, and the main process kept track of everything. This commit flips
things around so the renderer holds on to *everything*, and the main
process only communicates changes to translation sources. This enables
instantaneous translation switching (no need to wait for an IPC call!)
and opens up possibilities such as proper localisation inheritance and
user overrides.